Extract App Hosting YAML configuration loading into a common method (#10049)
* Extract App Hosting YAML configuration loading into a shared utility function and add environment variable splitting.

By extracting this code, it's possible to re-use the logic in other parts of our cli code.

* Add test verification

* remove unneccesary options/opts

* Get rid of unused SplitEnvVars code (for now)

* Fix linter error (define dynamicDispatch earlier)
